AI-service/ml-server/requirements.txt
# FastAPI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets>=12.0
wsproto>=1.2
python-multipart==0.0.6

# OCR
pytesseract==0.3.10
easyocr==1.7.1
# Pillow: use 10.4+ to avoid KeyError on Python 3.14 source build
Pillow>=10.2.0,<11

# ML
scikit-learn>=1.5.0
pandas>=2.2.0
# ChromaDB uses np.float_ removed in NumPy 2.0 → keep numpy 1.x
numpy>=1.26.3,<2
joblib>=1.3.2

# NLP (for text classification)
# konlpy==0.6.0  # disabled: needs C++ compiler on Python 3.14
# To install: install Microsoft C++ Build Tools then pip install konlpy
# Or: pip install JPype1>=1.5.0 (newer wheels may support Python 3.14)

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3

# Database & Vector Search
psycopg2-binary==2.9.9
pgvector==0.2.4
chromadb>=0.4.0

# LLM & Embedding
google-generativeai==0.3.2
google-genai>=1.0.0
httpx==0.26.0

# LangChain (Intent fallback 등)
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
